+++ global/util.gd
@@ -84,10 +84,14 @@ func is_shift_pressed() -> bool:
     return Input.is_key_pressed(KEY_SHIFT)
 
 func get_icon_for_color(icon_color: Color) -> Texture2D:
-    var icon_size: = ANESettings.MENU_ICON_SIZE
-    var img: = Image.create(icon_size, icon_size, false, Image.FORMAT_RGB8)
-    img.fill(icon_color)
-    return ImageTexture.create_from_image(img)
+    #var icon_size: = ANESettings.MENU_ICON_SIZE
+    #var img: = Image.create(icon_size, icon_size, false, Image.FORMAT_RGB8)
+    #img.fill(icon_color)
+    #return ImageTexture.create_from_image(img)
+    var color_blob: = preload("res://ui/assets/color_blob_icon.tres")
+    color_blob = color_blob.duplicate()
+    color_blob.color_map = { Color.BLACK: icon_color }
+    return color_blob
 
 func out_connections(conn_infos: Array[Dictionary], graph_node_name: String, only_at_port: int = -1) -> Array[Dictionary]:
     var out_conn_infos: Array[Dictionary] = []